Traktor: Linux compile fix.
[git-p4: depot-paths = "//depot/traktor/": change = 21321]

+++ code/Model/Operations/UnwrapUV.cpp
@@ -169,7 +169,7 @@ bool UnwrapUV::apply(Model& model) const
 			if (ln > mln)
 			{
 				mln = ln;
-				angle = std::atan2f(edge.y, edge.x);
+				angle = std::atan2(edge.y, edge.x);
 			}
 		}
